moved bazooka functions to a single file
