+++ core/bazooka.go
@@ -1,9 +1,14 @@
 package core
 
 import (
+	"context"
+	"errors"
+	"math/big"
 	"strings"
 
+	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/rpc"
+	"github.com/maticnetwork/bor/accounts/abi/bind"
 
 	"github.com/BOPR/common"
 	"github.com/BOPR/config"
@@ -104,4 +109,259 @@ func NewPreLoadedBazooka() (bazooka Bazooka, err error) {
 	bazooka.log = common.Logger.With("module", "bazooka")
 
 	return bazooka, nil
+}
+
+// get main chain block header
+func (b *Bazooka) GetMainChainBlock(blockNum *big.Int) (header *ethTypes.Header, err error) {
+	latestBlock, err := b.EthClient.HeaderByNumber(context.Background(), blockNum)
+	if err != nil {
+		return
+	}
+	return latestBlock, nil
+}
+
+// TotalBatches returns the total number of batches that have been submitted on chain
+func (b *Bazooka) TotalBatches() (uint64, error) {
+	totalBatches, err := b.RollupContract.NumOfBatchesSubmitted(nil)
+	if err != nil {
+		return 0, err
+	}
+	return totalBatches.Uint64(), nil
+}
+
+func (b *Bazooka) FetchBalanceTreeRoot() (ByteArray, error) {
+	root, err := b.RollupContract.GetLatestBalanceTreeRoot(nil)
+	if err != nil {
+		return ByteArray{}, err
+	}
+	return root, nil
+}
+
+func (b *Bazooka) FetchBatchInputData(txHash ethCmn.Hash) (txs [][]byte, err error) {
+	tx, isPending, err := b.EthClient.TransactionByHash(context.Background(), txHash)
+	if err != nil {
+		b.log.Error("Cannot fetch transaction from hash", "Error", err)
+		return
+	}
+
+	if isPending {
+		err := errors.New("Transaction is pending")
+		b.log.Error("Transaction is still pending, cannot process", "Error", err)
+		return txs, err
+	}
+
+	payload := tx.Data()
+	decodedPayload := payload[4:]
+	inputDataMap := make(map[string]interface{})
+	method := b.ContractABI[common.ROLLUP_CONTRACT_KEY].Methods["submitBatch"]
+	err = method.Inputs.UnpackIntoMap(inputDataMap, decodedPayload)
+	if err != nil {
+		b.log.Error("Error unpacking payload", "Error", err)
+		return
+	}
+
+	return GetTxsFromInput(inputDataMap), nil
+}
+
+// ProcessTx calls the ProcessTx function on the contract to verify the tx
+// returns the updated accounts and the new balance root
+func (b *Bazooka) ProcessTx(balanceTreeRoot, accountTreeRoot ByteArray, tx Tx, fromMerkleProof, toMerkleProof AccountMerkleProof, pdaProof PDAMerkleProof) (newBalanceRoot ByteArray, from, to []byte, err error) {
+	txABIVersion := tx.ToABIVersion(int64(tx.From), int64(tx.To))
+	opts := bind.CallOpts{From: config.OperatorAddress}
+	typesAccountProofs := rollup.TypesAccountProofs{From: fromMerkleProof.ToABIVersion(), To: toMerkleProof.ToABIVersion()}
+	updatedRoot, newFromAccount, newToAccount, errCode, IsValidTx, err := b.RollupContract.ProcessTx(&opts,
+		balanceTreeRoot,
+		accountTreeRoot,
+		txABIVersion,
+		pdaProof.ToABIVersion(),
+		typesAccountProofs,
+	)
+	if err != nil {
+		return
+	}
+
+	b.log.Info("Processed transaction", "IsSuccess", IsValidTx, "newRoot", updatedRoot)
+
+	if !IsValidTx {
+		b.log.Error("Invalid transaction", "error_code", errCode)
+		return newBalanceRoot, from, to, errors.New("Tx is invalid")
+	}
+	newBalanceRoot = BytesToByteArray(updatedRoot[:])
+	return newBalanceRoot, newFromAccount, newToAccount, nil
+}
+
+func (b *Bazooka) ApplyTransferTx(account AccountMerkleProof, tx Tx) ([]byte, ByteArray, error) {
+	txABIVersion := tx.ToABIVersion(int64(tx.From), int64(tx.To))
+	updatedAccountBytes, updatedRoot, err := b.RollupContract.ApplyTx(nil, account.ToABIVersion(), txABIVersion)
+	if err != nil {
+		return updatedAccountBytes, updatedRoot, err
+	}
+
+	return updatedAccountBytes, updatedRoot, nil
+}
+
+// func (b *Bazooka) CompressTx(tx Tx) ([]byte, ByteArray, error) {
+// 	txABIVersion := tx.ToABIVersion(int64(tx.From), int64(tx.To))
+
+// 	b.RollupUtils.CompressTx(nil, txABIVersion)
+// }
+
+// func (b *Bazooka) EncodeTx(from, to, token, amount int64) ([]byte, error) {
+// 	return b.RollupUtils.BytesFromTxDeconstructed(nil, big.NewInt(from), big.NewInt(to), big.NewInt(token), big.NewInt(amount))
+// }
+
+// func (b *Bazooka) DecodeTx(txBytes []byte) (rolluputils.TypesTransaction, error) {
+// 	tx, err := b.RollupUtils.TxFromBytes(nil, txBytes)
+// 	return tx, err
+// }
+
+func (b *Bazooka) EncodeAccount(id, balance, nonce, token int64) (accountBytes []byte, err error) {
+	accountBytes, err = b.RollupUtils.BytesFromAccountDeconstructed(nil, big.NewInt(id), big.NewInt(balance), big.NewInt(nonce), big.NewInt(token))
+	if err != nil {
+		return
+	}
+	return accountBytes, nil
+}
+
+func (b *Bazooka) DecodeAccount(accountBytes []byte) (ID, balance, nonce, token uint64, err error) {
+	account, err := b.RollupUtils.AccountFromBytes(nil, accountBytes)
+	if err != nil {
+		return
+	}
+
+	return account.ID.Uint64(), account.Balance.Uint64(), account.Nonce.Uint64(), account.TokenType.Uint64(), nil
+}
+
+func (b *Bazooka) FireDepositFinalisation(TBreplaced UserAccount, siblings []UserAccount, subTreeHeight uint64) error {
+	b.log.Info(
+		"Attempting to finalise deposits",
+		"NodeToBeReplaced",
+		TBreplaced.String(),
+		"NumberOfSiblings",
+		len(siblings),
+		"atDepth",
+		subTreeHeight,
+	)
+
+	// TODO check latest batch on-chain and if we need to push new batch
+
+	depositSubTreeHeight := big.NewInt(0)
+	depositSubTreeHeight.SetUint64(subTreeHeight)
+	var siblingData [][32]byte
+	for _, sibling := range siblings {
+		data, err := HexToByteArray(sibling.Hash)
+		if err != nil {
+			return err
+		}
+		siblingData = append(siblingData, data)
+	}
+	accountProof := rollup.TypesAccountMerkleProof{}
+	accountProof.AccountIP.PathToAccount = StringToBigInt(TBreplaced.Path)
+	accountProof.AccountIP.Account = TBreplaced.ToABIAccount()
+	accountProof.Siblings = siblingData
+	data, err := b.ContractABI[common.ROLLUP_CONTRACT_KEY].Pack("finaliseDepositsAndSubmitBatch", depositSubTreeHeight, accountProof)
+	if err != nil {
+		return err
+	}
+
+	rollupAddress := ethCmn.HexToAddress(config.GlobalCfg.RollupAddress)
+	stakeAmount := big.NewInt(0)
+	stakeAmount.SetString("32000000000000000000", 10)
+
+	// generate call msg
+	callMsg := ethereum.CallMsg{
+		To:    &rollupAddress,
+		Data:  data,
+		Value: stakeAmount,
+	}
+
+	auth, err := b.GenerateAuthObj(b.EthClient, callMsg)
+	if err != nil {
+		return err
+	}
+	b.log.Info("Broadcasting deposit finalisation transaction")
+	tx, err := b.RollupContract.FinaliseDepositsAndSubmitBatch(auth, depositSubTreeHeight, accountProof)
+	if err != nil {
+		return err
+	}
+	b.log.Info("Deposits successfully finalized!", "TxHash", tx.Hash())
+	return nil
+}
+
+// SubmitBatch submits the batch on chain with updated root and compressed transactions
+func (b *Bazooka) SubmitBatch(updatedRoot ByteArray, txs []Tx) error {
+	b.log.Info(
+		"Attempting to submit a new batch",
+		"UpdatedRoot",
+		updatedRoot.String(),
+		"txs",
+		len(txs),
+	)
+
+	var compressedTxs [][]byte
+	for _, tx := range txs {
+		compressedTx, err := tx.Compress()
+		if err != nil {
+			return err
+		}
+		compressedTxs = append(compressedTxs, compressedTx)
+	}
+
+	data, err := b.ContractABI[common.ROLLUP_CONTRACT_KEY].Pack("submitBatch", compressedTxs, updatedRoot)
+	if err != nil {
+		return err
+	}
+
+	rollupAddress := ethCmn.HexToAddress(config.GlobalCfg.RollupAddress)
+
+	// generate call msg
+	callMsg := ethereum.CallMsg{
+		To:   &rollupAddress,
+		Data: data,
+	}
+
+	auth, err := b.GenerateAuthObj(b.EthClient, callMsg)
+	if err != nil {
+		return err
+	}
+
+	tx, err := b.RollupContract.SubmitBatch(auth, compressedTxs, updatedRoot)
+	if err != nil {
+		return err
+	}
+	b.log.Info("Sent a new batch!", "txHash", tx.Hash().String())
+	return nil
+}
+
+func GetTxsFromInput(input map[string]interface{}) (txs [][]byte) {
+	return input["_txs"].([][]byte)
+}
+
+func (b *Bazooka) GenerateAuthObj(client *ethclient.Client, callMsg ethereum.CallMsg) (auth *bind.TransactOpts, err error) {
+	// from address
+	fromAddress := config.OperatorAddress
+
+	// fetch gas price
+	gasprice, err := client.SuggestGasPrice(context.Background())
+	if err != nil {
+		return
+	}
+
+	// fetch nonce
+	nonce, err := client.PendingNonceAt(context.Background(), fromAddress)
+	if err != nil {
+		return
+	}
+
+	// fetch gas limit
+	callMsg.From = fromAddress
+	gasLimit, err := client.EstimateGas(context.Background(), callMsg)
+
+	// create auth
+	auth = bind.NewKeyedTransactor(config.OperatorKey)
+	auth.GasPrice = gasprice
+	auth.Nonce = big.NewInt(int64(nonce))
+	auth.GasLimit = uint64(gasLimit) // uint64(gasLimit)
+
+	return
 }
