+++ core/utils.go
@@ -1,7 +1,6 @@
 package core
 
 import (
-	"context"
 	"fmt"
 	"math/big"
 	"strconv"
@@ -9,11 +8,7 @@ import (
 	"unicode/utf8"
 
 	"github.com/BOPR/common"
-	"github.com/BOPR/config"
-	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/accounts/abi"
-	"github.com/ethereum/go-ethereum/accounts/abi/bind"
-	"github.com/ethereum/go-ethereum/ethclient"
 	"github.com/willf/pad"
 )
 
@@ -166,37 +161,4 @@ func SolidityPathToNodePath(path uint64, depth uint64) (string, error) {
 	}
 	generatedPath := strings.Join([]string{string(pathToNode), pathWithoutPrefix}, "")
 	return generatedPath, nil
-}
-
-func GetTxsFromInput(input map[string]interface{}) (txs [][]byte) {
-	return input["_txs"].([][]byte)
-}
-
-func (b *Bazooka) GenerateAuthObj(client *ethclient.Client, callMsg ethereum.CallMsg) (auth *bind.TransactOpts, err error) {
-	// from address
-	fromAddress := config.OperatorAddress
-
-	// fetch gas price
-	gasprice, err := client.SuggestGasPrice(context.Background())
-	if err != nil {
-		return
-	}
-
-	// fetch nonce
-	nonce, err := client.PendingNonceAt(context.Background(), fromAddress)
-	if err != nil {
-		return
-	}
-
-	// fetch gas limit
-	callMsg.From = fromAddress
-	gasLimit, err := client.EstimateGas(context.Background(), callMsg)
-
-	// create auth
-	auth = bind.NewKeyedTransactor(config.OperatorKey)
-	auth.GasPrice = gasprice
-	auth.Nonce = big.NewInt(int64(nonce))
-	auth.GasLimit = uint64(gasLimit) // uint64(gasLimit)
-
-	return
 }
